+++ changelog.md
@@ -5,6 +5,9 @@ Version History
 -----------
 Here you'll find information on key changes, bug fixes and improvements made by version.
 
+### What's new in version 1809.2?
+* App Service SQL DB Cleanup for reruns
+
 ### What's new in version 1809.1?
 * Support for Database RPs 1.1.30.0
 * Support for App Service 1.4 (Update 4)