+++ changelog.md
@@ -7,6 +7,7 @@ Here you'll find information on key changes, bug fixes and improvements made by 
 
 ### What's new in version 1809.2?
 * App Service SQL DB Cleanup for reruns
+* Cleans up App Service Resource Group in case of previous run failure - ensures fresh next attempt
 
 ### What's new in version 1809.1?
 * Support for Database RPs 1.1.30.0