+++ changelog.md
@@ -5,6 +5,9 @@ Version History
 -----------
 Here you'll find information on key changes, bug fixes and improvements made by version.
 
+### What's new in version 1811?
+* Bug fixes
+
 ### What's new in version 1809.3?
 * Adjusted VM sizes for Resource Providers to use less resources
 * Added host memory check to avoid running out of memory