+++ changelog.md
@@ -6,6 +6,7 @@ Version History
 Here you'll find information on key changes, bug fixes and improvements made by version.
 
 ### What's new in version 1811?
+* Updated to support ASDK 1.1811.0.101
 * Bug fixes
 
 ### What's new in version 1809.3?